uploading to s3: grouping files by dir name; writing JSONS to files

+++ workflows/collect_save_and_upload_data/solids/uploading_files_to_s3.py
@@ -31,21 +31,26 @@ def uploading_files_to_s3(context):
 
     pwd = os.getcwd()
     path_to_files = os.path.join(pwd, UNZIPED_FILES)
-    file_names = [file_name for file_name in os.listdir(path_to_files) if
-                  os.path.isfile(os.path.join(path_to_files, file_name))]
-    for file_name in file_names:
-        path_to_file = os.path.join(pwd, f'{UNZIPED_FILES}/{file_name}')
-        key = f"{DUMMY_FILES_SUB_KEY}/{file_name}"
 
-        s3_resource.Bucket(bucket_name).put_object(
-            Key=key,
-            Body=open(path_to_file, 'rb')
-        )
+    for dir_name in os.listdir(path_to_files):
+        if os.path.isdir(os.path.isfile(os.path.join(path_to_files, dir_name))):
+            dir_full_path = os.path.join(path_to_files, dir_name)
+            file_names = [file_name for file_name in os.listdir(dir_full_path) if
+                          os.path.isfile(os.path.join(dir_full_path, file_name))]
+
+            for file_name in file_names:
+                path_to_file = os.path.join(pwd, f'{UNZIPED_FILES}/{dir_name}/{file_name}')
+                key = f"{DUMMY_FILES_SUB_KEY}/{dir_name}/{file_name}"
+
+                s3_resource.Bucket(bucket_name).put_object(
+                    Key=key,
+                    Body=open(path_to_file, 'rb')
+                )
     # return keys of files, where they were uploaded.
     # Because after this step we will use just Keys of s3 to re-download files to modify them
     for key in s3_client.list_objects(Bucket=bucket_name)['Contents']:
         # keys, which ended with '/' are dirs, we need just files
         # files, which we want to download from s3
-        if not key['Key'].endswith('/') and f'{DUMMY_FILES_SUB_KEY}/' in key['Key']:
+        if key['Key'].endswith('/') and f'{DUMMY_FILES_SUB_KEY}/' in key['Key']:
             yield DynamicOutput(value=key['Key'],
                                 mapping_key=generate_mapping_key())
